Change way of signing packages
Re-purpose the `scripts/publish.py` script into `scripts/sign.py`, and
use it only for signing, since we already have a verify-only script
(`scripts/check.py`).
